requirements.txt
fastapi
uvicorn
scikit-learn
joblib
python-dotenv
streamlit
pandas
numpy
scikit-learn
imaplib2
pytest
flake8
